requirements.txt
matplotlib==3.10.3
numpy==2.3.0
Pillow==11.2.1
tensorflow==2.19.0
